hw03-main/hw03/requirements.txt
opencv-python==4.10.0
streamlit==1.32.2
pillow==10.2.0
numpy==1.26.4